+++ src/Routes/Routes.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import Login from '../pages/Login';
-import Home from '../pages/Home';
+import Login from '../Pages/Login';
+import Home from '../Pages/Home';
 
 const Routes = () => {
   return (
